Arrange for /favicon.ico to at least result in a redirect
I can't arrange for /favicon.ico to be served by the Apache that fronts
Gunicorn (the config forwards all requests to Gunicorn). I could write a view
that reads and sends the content of favicon.ico, but all I really want is
for /favicon.ico not to trigger a 404. So this edit arranges for a redirect
to the corresponding static asset to be sent. This may not be good enough for
some browsers, but I don't actually care!