Fix missing authentication on account/edit
The '@' was missing from the '@login_required()' decorator on
account_edit(), with the result that it wasn't. No serious harm done
becasue the attempt to access the 'email' attribute of the resulting
AnonymousUser through an error.

This was a particular problem because account/edit is linked from the
bottom of all the generated emails but is likely to be invoked when
not already authenticated.