Change handling of cancel, (un)volunteer, and accout edit pages
Rather than falling through to the confirmation page with error
messages and the submit button disabled if an action isn't
allowed, redirect to the details page event-details, account).

As well as (probably) making more sense, this avoids any
problems with the confirmation page displaying information
(such as current volunteers) that the current user may not
be allowed to see (e.g. if the problem is that they are
not event owner).

Closes #60