Add helper and organiser guidelines, rework nav bar
Add the current draft guidelines for helpers and organisers.

Link to helper guidelines from the 'volunteer' page, and to
the organiser guidelines from 'create event' page.

Turn the existing 'About' menu item into a 'Help' drop-down
that links to the existing About and the two sets of guidelines.

Turn the existing 'userid - Logout - Account' sequence at the right
of the nav bar into a drop-down to save space; revert the nav bar
expand point to md because the bar is now narrower and this means
it's displayed in full on a portrait 9.7" iPad.

Make the nav bar sticky-top so it stays on the screen (could just
be fixed-top onve the 'Demo oly' banner goes away).