Fixup scheme and protocol in emails
Arrange to pass scheme and domain into email templates. Done this using
parameters in settings because the expected hosting for this app doesn't
seem to pass scheme correctly from the Apache frontend into the
Gunicorn wsgi handler.

Also sub-classed the supplied PasswordResetView to pass our idea of scheme
to the email template.