Rework volunteering to keep volounteer records
Previously, withdrawing or declining offers of help was handled
by deleting the corresponding volunteer record. But this leaves no
record of what happened which could be a problem for user support
and means we can't derive statistics. And it's at odds with how
we handle events.

This edit adds 'withdrawn' and 'declined' fields to the volunteer
record (and removed the unused 'cancelled') and re-works the logic
so that on;y records with both NULL count as current. Turns out
to be quite an extensive edit.

Also fixed a bug that triggered a bogus clash error when editing
events without changing that times or location.