For the short event timestamp, show year in not current
The short timestamp used to show the year for events in
December and January, but not otherwise. This isn't helpful
in the middle of a long list of events. Switch to showing
year for events in any year except the current one. This
should mean that year will be omited in most common cases
while still appearing is needed. Omiting year saves display
space on narrow screens.